Automatically display the help text for main command
This means the applications will no longer need to push `--help` onto
`ARGV` if called without any arguments.

The error texted is also tweaked not to display the main error message
in this case. This is because it isn't really an error, even though it
is handled as an error