Sort the subcommand as an array not a hash
When a hash is sorted, it always sorts by the key. This means it is only
possible to sort in alphabetical order.

Instead the command hash is casted to an array before sorting. This
means the spaceship operator is used `<=>` to preform the comparison.

The operator is added using a patch. This will be used to assign
priorities to the commands eventually