Remove the use of a singleton runner
Currently the singleton runner means the CLI can only ever
work on the ARGV of the process. This a pain as it should be
selectable by the end user

This is the first step of rectifying the issue by moving the
instance onto the Builder. Eventually the Runner will be made
dynamically on demand