Make the error handler program aware
Version 1 of Commander use to colourize and include the program name
in its error messages. This was lost in the transition to Version 2.

To support this again, an ErrorHandler struct has been created. This
allows it to generate error messages with the program_name.

It also more clearly defines the work flow triggering the tracing. The
ErrorHandler as the ability to parse for the `--trace` flag without
the use of SLOP. This allows `--trace` to work with internal errors

Conveniently it does allow the main error handler to start before
parsing for --trace

TL;DR There is a new ErrorHandler that pretty colourizes the errors
      + other --trace stuff